services/opensearch_puller/requirements.txt
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
requests==2.32.3
cryptography==42.0.8
